+++ docs/src/advanced-usage/manually-observe-state/observe-specific-state-vars/specific-manual-component.js
@@ -61,7 +61,7 @@ export class SpecificManualComponent extends demoComponentStyle(LitElement) {
 
     handleObserveCounter1ButtonClick() {
         this.counter1Observer = () => this.requestUpdate();
-        demoState.addObserver(this.counter1Observer, 'counter1');
+        demoState.addObserver(this.counter1Observer, ['counter1']);
         this.observingCounter1 = true;
     }
 
@@ -72,7 +72,7 @@ export class SpecificManualComponent extends demoComponentStyle(LitElement) {
 
     handleObserveCounter2ButtonClick() {
         this.counter2Observer = () => this.requestUpdate();
-        demoState.addObserver(this.counter2Observer, 'counter2');
+        demoState.addObserver(this.counter2Observer, ['counter2']);
         this.observingCounter2 = true;
     }